fix: unmodifiableList를 정렬할 수 없는 문제 해결
정렬이 안되어 생성자를 초기화할 수 없었던 문제를 해결했습니다.

+++ src/main/java/lotto/app/dto/LottoResultResponseDto.java
@@ -1,5 +1,5 @@
 package lotto.app.dto;
 
-public record LottoResultResponseDto(int correctNumber, int prize, int amount) {
+public record LottoResultResponseDto(int correctNumber, Long prize, int amount) {
 
 }
